General text cleanliness and other miscellaneous
