+++ src/main/res/layout/activity_seasonal_per_season.xml
@@ -11,12 +11,10 @@
         android:id="@+id/pageTitle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="173dp"
-        android:layout_marginTop="320dp"
-        android:layout_marginEnd="173dp"
-        android:layout_marginBottom="320dp"
-        android:text="Among Us"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginStart="182dp"
+        android:layout_marginTop="32dp"
+        android:layout_marginEnd="183dp"
+        android:text="Season"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
